Move gc_protector.py to it righteous place
