+++ pythonscript/embedded/godot/hazmat/ffi/script.py
@@ -4,7 +4,7 @@ import traceback
 from pythonscriptcffi import ffi, lib
 
 from godot.hazmat.base import BaseObject, get_exposed_class_per_module
-from godot.hazmat.ffi.gc_protector import connect_handle
+from godot.hazmat.gc_protector import connect_handle
 from godot.hazmat.tools import (
     godot_string_to_pyobj,
     godot_string_from_pyobj,
