+++ pythonscript/embedded/godot/hazmat/gc_protector.py
